+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-17T05:59:07.739Z",
+    "2026-02-17T11:54:17.102Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {