+++ packages/orq-rc/RELEASES.md
@@ -6869,3 +6869,13 @@ Based on:
 - [typescript v4.4.0-rc.4] packages/orq-rc
 ### Releases
 - [NPM v4.4.0-rc.4] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.4 - packages/orq-rc
+
+## 2026-02-17 11:57:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.721.0 (2.828.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.4.0-rc.5] packages/orq-rc
+### Releases
+- [NPM v4.4.0-rc.5] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.5 - packages/orq-rc